pyproject.toml
[project]
name = "job-agent"
version = "0.1.0"
description = "A multi-agent system to streamline job application process"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.31.0",
    "streamlit-option-menu>=0.3.12",
    "openai>=1.12.0",
    "agno>=0.1.0",
    "PyPDF2>=3.0.1",
    "beautifulsoup4>=4.12.3",
    "requests>=2.31.0",
    "selenium>=4.18.1",
    "webdriver-manager>=4.0.1",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.1",
    "pymongo>=4.13.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "pages", "utils"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ["py39"]
